+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Improved `LimitedVector::insert` to deal better with conversions and complex types.
 * Improved `LimitedMap`, `LimitedOrdered`, `LimitedSet` and `LimitedVector` ability to handle conversions.
+* Added diff tooling options `ignore_all_space` and `ignore_consecutive_space`.
+* Changed diff tooling options `ignore_space_change` to only ignore trailing space to match `git diff`.
 
 # 0.2.27